Improved SoundBackend

+++ src/backends/AnalyticsBackend.cpp
@@ -111,14 +111,14 @@ void ABB::AnalyticsBackend::reset() {
 
 
 float ABB::AnalyticsBackend::getStackSizeBuf(void* data, int ind){
-    RingBuffer<uint32_t>* stackSizeBufPtr = (RingBuffer<uint32_t>*)data;
+    RingBuffer<uint16_t>* stackSizeBufPtr = (RingBuffer<uint16_t>*)data;
     if((size_t)ind >= stackSizeBufPtr->size()){
         return 0;
     }
     return stackSizeBufPtr->get(ind);
 }
 float ABB::AnalyticsBackend::getSleepCycsBuf(void* data, int ind){
-    RingBuffer<uint32_t>* sleepCycsBufPtr = (RingBuffer<uint32_t>*)data;
+    RingBuffer<uint64_t>* sleepCycsBufPtr = (RingBuffer<uint64_t>*)data;
     if((size_t)ind >= sleepCycsBufPtr->size()){
         return 0;
     }
